pyproject.toml
[tool.poetry]
name = "apac-coe-technical-test-fall-2023"
version = "0.1.0"
description = "This is my attempt for APAC COE Technical Test (Full-Stack)"
authors = ["Richard Chan <manyanchanmanyan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.105.0"
sqlalchemy = "^2.0.23"
pymysql = "^1.1.0"
pyjwt = "^2.8.0"
dash = "^2.14.2"
pandas = "^2.1.4"
python-dotenv = "^1.0.0"
uvicorn = "^0.24.0.post1"
httpx = "^0.25.2"
python-multipart = "^0.0.6"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
ruff = "^0.1.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
